+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `client` option for non-promise client
+
 ## [4.2.2] - 2021-03-02
 
 ### Fixed